+++ docs/CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Next
 
+* Package `context`
+  * New `VariableWithValueGraph` to create/get a variable with value set to the graph value (Node).
+  * New `IterVariables` and `IterVariablesInScope` that use the new go 1.23 iterators.
 * New directory `ui` with various front-ends for displaying training progress, plots, etc.
   * **BREAKING CHANGE**: Refactored all UI tools under `ui` directory. It only requires changing the import, the APIs are not changed.
   * New package `fyneui`, a window based training UI built using Fyne.io (EXPERIMENTAL)
@@ -12,6 +15,7 @@
 * New package `cosineschedule`, refactored from `optimizers` package.
   * Added handling negative values for the hyperparameter `cosine_schedule_steps`: they set the period of the cosine schedule
     as fractions of the total number of steps being trained.
+
 
 ## v0.13.0 - 2024/10/07